[LCR-205] refactor: controller 패키지명을 externalApi로 변경
